+++ astropy/coordinates/earth.py
@@ -2,6 +2,7 @@
 from __future__ import absolute_import, division, print_function
 
 from warnings import warn
+import collections
 import socket
 import json
 
@@ -16,7 +17,7 @@ from ..utils.compat.numpy import broadcast_to
 from .angles import Longitude, Latitude
 from .representation import CartesianRepresentation
 from .errors import UnknownSiteException
-from ..utils import data
+from ..utils import data, deprecated
 
 try:
     # Not guaranteed available at setup time.
@@ -26,6 +27,8 @@ except ImportError:
         raise
 
 __all__ = ['EarthLocation']
+
+GeodeticLocation = collections.namedtuple('GeodeticLocation', ['lon', 'lat', 'height'])
 
 # Available ellipsoids (defined in erfam.h, with numbers exposed in erfa).
 ELLIPSOIDS = ('WGS84', 'GRS80', 'WGS72')
@@ -475,19 +478,32 @@ class EarthLocation(u.Quantity):
         ellipsoid = _check_ellipsoid(ellipsoid, default=self.ellipsoid)
         self_array = self.to(u.meter).view(self._array_dtype, np.ndarray)
         lon, lat, height = erfa.gc2gd(getattr(erfa, ellipsoid), self_array)
-        return (Longitude(lon * u.radian, u.degree,
-                          wrap_angle=180.*u.degree, copy=False),
-                Latitude(lat * u.radian, u.degree, copy=False),
-                u.Quantity(height * u.meter, self.unit, copy=False))
+        return GeodeticLocation(
+            Longitude(lon * u.radian, u.degree,
+                      wrap_angle=180.*u.degree, copy=False),
+            Latitude(lat * u.radian, u.degree, copy=False),
+            u.Quantity(height * u.meter, self.unit, copy=False))
 
     @property
+    @deprecated('2.0', alternative='`lon`', obj_type='property')
     def longitude(self):
         """Longitude of the location, for the default ellipsoid."""
         return self.geodetic[0]
 
     @property
+    def lon(self):
+        """Longitude of the location, for the default ellipsoid."""
+        return self.geodetic[0]
+
+    @property
+    @deprecated('2.0', alternative='`lat`', obj_type='property')
     def latitude(self):
         """Latitude of the location, for the default ellipsoid."""
+        return self.geodetic[1]
+
+    @property
+    def lat(self):
+        """Longitude of the location, for the default ellipsoid."""
         return self.geodetic[1]
 
     @property